Add controller actions for items search requests
