+++ app/models/merchant.rb
@@ -6,28 +6,4 @@ class Merchant < ApplicationRecord
   has_many :items, dependent: :destroy
   has_many :invoices, dependent: :destroy
   has_many :customers, through: :invoices
-
-  def self.find_all_by_name(name)
-    where('lower(name) LIKE ?', "%#{name.downcase}%")
-  end
-
-  def self.find_all_by_created_at(created_at)
-    where(created_at: created_at)
-  end
-
-  def self.find_all_by_updated_at(updated_at)
-    where(updated_at: updated_at)
-  end
-
-  def self.find_by_name(name)
-    where('lower(name) LIKE ?', "%#{name.downcase}%").limit(1).first
-  end
-
-  def self.find_by_created_at(created_at)
-    where(created_at: created_at).limit(1).first
-  end
-
-  def self.find_by_updated_at(updated_at)
-    where(updated_at: updated_at).limit(1).first
-  end
 end
