+++ config/routes.rb
@@ -1,6 +1,11 @@
 Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
+      namespace :items do
+        get '/find_all', to: 'search#index'
+        get '/find', to: 'search#show'
+      end
+
       namespace :merchants do
         get '/find_all', to: 'search#index'
         get '/find', to: 'search#show'
